refactor: 팀 리뷰 수정_1 (#99)

+++ src/main/java/com/mr/domain/playing/repository/PlayingRepository.java
@@ -101,11 +101,11 @@ public interface PlayingRepository extends JpaRepository<Playing, Long> {
             select coalesce(sum(p.durationSec), 0) from Playing p
             where p.user.userId = :userId
               and p.status = :status
-              and p.deletedAt is null
+              and p.deletedAt is null 
               and p.endedAt >= :since
               and p.id != :excludePlayingId
             """)
-    Integer sumDurationSecExcludeCurrent(
+    Long sumDurationSecExcludeCurrent(
             @Param("userId") Long userId,
             @Param("status") PlayingStatus status,
             @Param("since") LocalDateTime since,
